Using h2 database in local development, modify spring security settings for connect h2 database console
