+++ src/main/java/com/sullung2yo/seatcatcher/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.sullung2yo.seatcatcher.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
