+++ src/main/java/com/sullung2yo/seatcatcher/config/ProjectConfig.java
@@ -19,38 +19,4 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 @ComponentScan(basePackages = "com.sullung2yo.seatcatcher")
 @RequiredArgsConstructor
 public class ProjectConfig {
-
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .csrf(AbstractHttpConfigurer::disable) // CSRF 보안 설정을 비활성화
-                .cors(AbstractHttpConfigurer::disable) // CORS 설정 비활성화 해서 모든 도메인에서 요청을 받을 수 있게 설정
-                .sessionManagement( // 세션 관리 설정 -> REST API 서버이기 때문에 세션을 사용하지 않으므로 STATELESS로 설정
-                        session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                )
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/user/authenticate").permitAll() // JWT 토큰 발급 API는 누구나 접근 가능
-                        .requestMatchers("/health").permitAll() // 헬스체크 용 API 엔드포인트
-                        .anyRequest().authenticated()
-                )
-                .exceptionHandling(exception -> exception
-                        .authenticationEntryPoint((request, response, authException) -> {
-                            // 인증 실패 시 (토큰 없거나 잘못됨) -> 401
-                            response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthorized");
-                        })
-                        .accessDeniedHandler((request, response, accessDeniedException) -> {
-                            // 인가 실패(denyAll, 권한 부족) -> 403
-                            response.sendError(HttpServletResponse.SC_FORBIDDEN, "Access Denied");
-                        })
-                )
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class); // JWT 토큰 검증 필터 추가
-        return http.build();
-    }
 }
